chore(tangle-dapp): Extract input to a separate component
