+++ apps/tangle-dapp/app/restake/delegate/page.tsx
@@ -3,21 +3,14 @@
 import isDefined from '@webb-tools/dapp-types/utils/isDefined';
 import Button from '@webb-tools/webb-ui-components/components/buttons/Button';
 import FeeDetails from '@webb-tools/webb-ui-components/components/FeeDetails';
-import type { TextFieldInputProps } from '@webb-tools/webb-ui-components/components/TextField/types';
-import { TransactionInputCard } from '@webb-tools/webb-ui-components/components/TransactionInputCard';
 import keys from 'lodash/keys';
 import { useCallback, useEffect, useMemo } from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
-import { formatUnits } from 'viem';
 
-import { useRestakeContext } from '../../../context/RestakeContext';
-import useRestakeConsts from '../../../data/restake/useRestakeConsts';
 import useRestakeDelegatorInfo from '../../../data/restake/useRestakeDelegatorInfo';
 import type { DelegationFormFields } from '../../../types/restake';
-import decimalsToStep from '../../../utils/decimalsToStep';
-import { getAmountValidation } from '../../../utils/getAmountValidation';
-import ErrorMessage from '../ErrorMessage';
 import RestakeTabs from '../RestakeTabs';
+import DelegationInput from './DelegationInput';
 
 export default function DelegatePage() {
   const {
@@ -30,30 +23,29 @@ export default function DelegatePage() {
     mode: 'onBlur',
   });
 
-  const { assetMap } = useRestakeContext();
-  const { delegatorInfo } = useRestakeDelegatorInfo();
-  const { minDelegateAmount } = useRestakeConsts();
-
-  const depositedAssets = useMemo(() => {
-    if (delegatorInfo === null) {
-      return [];
-    }
-
-    return keys(delegatorInfo.deposits).map((assetId) => assetMap[assetId]);
-  }, [assetMap, delegatorInfo]);
-
-  // Set the default assetId to the first assetId in the depositedAssets
-  const defaultAssetId = useMemo(
-    () => (depositedAssets.length > 0 ? depositedAssets[0].id : null),
-    [depositedAssets],
-  );
-
   // Register select fields on mount
   useEffect(() => {
     register('assetId', { required: 'Asset is required' });
     register('operatorAccountId', { required: 'Operator is required' });
   }, [register]);
 
+  const { delegatorInfo } = useRestakeDelegatorInfo();
+
+  // Set the default assetId to the first assetId in the depositedAssets
+  const defaultAssetId = useMemo(() => {
+    if (!isDefined(delegatorInfo)) {
+      return null;
+    }
+
+    const assetIds = keys(delegatorInfo.deposits);
+    if (assetIds.length === 0) {
+      return null;
+    }
+
+    return assetIds[0];
+  }, [delegatorInfo]);
+
+  // Set the default assetId to the first deposited assets
   useEffect(() => {
     if (defaultAssetId !== null) {
       setValue('assetId', defaultAssetId, {
@@ -62,73 +54,6 @@ export default function DelegatePage() {
       });
     }
   }, [defaultAssetId, setValue]);
-
-  const selectedAssetId = watch('assetId');
-
-  const selectedAsset = useMemo(
-    () => (selectedAssetId !== null ? assetMap[selectedAssetId] : null),
-    [assetMap, selectedAssetId],
-  );
-
-  const { max, maxFormatted } = useMemo(() => {
-    if (!isDefined(selectedAsset) || !isDefined(delegatorInfo)) {
-      return {};
-    }
-
-    const amountRaw = delegatorInfo.deposits[selectedAsset.id]?.amount;
-    if (!isDefined(amountRaw)) {
-      return {};
-    }
-
-    const maxFormatted = +formatUnits(amountRaw, selectedAsset.decimals);
-
-    return {
-      max: amountRaw,
-      maxFormatted,
-    };
-  }, [delegatorInfo, selectedAsset]);
-
-  const { min, minFormatted } = useMemo(() => {
-    if (!isDefined(minDelegateAmount) || !isDefined(selectedAsset)) {
-      return {};
-    }
-
-    return {
-      min: minDelegateAmount,
-      minFormatted: formatUnits(minDelegateAmount, selectedAsset.decimals),
-    };
-  }, [minDelegateAmount, selectedAsset]);
-
-  const customAmountProps = useMemo<TextFieldInputProps>(
-    () => {
-      const step = decimalsToStep(selectedAsset?.decimals);
-
-      return {
-        type: 'number',
-        step,
-        ...register('amount', {
-          required: 'Amount is required',
-          validate: getAmountValidation(
-            step,
-            minFormatted,
-            min,
-            max,
-            selectedAsset?.decimals,
-            selectedAsset?.symbol,
-          ),
-        }),
-      };
-    },
-    // prettier-ignore
-    [max, min, minFormatted, register, selectedAsset?.decimals, selectedAsset?.symbol],
-  );
-
-  const handleAmountChange = useCallback(
-    (amount: string) => {
-      setValue('amount', amount, { shouldDirty: true, shouldValidate: true });
-    },
-    [setValue],
-  );
 
   const onSubmit = useCallback<SubmitHandler<DelegationFormFields>>((data) => {
     console.log(data);
@@ -143,28 +68,13 @@ export default function DelegatePage() {
         <RestakeTabs />
 
         <div className="space-y-2">
-          <TransactionInputCard.Root
-            tokenSymbol={selectedAsset?.symbol}
-            maxAmount={maxFormatted}
-            onAmountChange={handleAmountChange}
-          >
-            <TransactionInputCard.Header>
-              <TransactionInputCard.ChainSelector placeholder="Select Operator" />
-              <TransactionInputCard.MaxAmountButton
-                accountType="note"
-                tooltipBody="Deposited Asset"
-              />
-            </TransactionInputCard.Header>
-
-            <TransactionInputCard.Body
-              tokenSelectorProps={{
-                placeHolder: 'Select Asset',
-              }}
-              customAmountProps={customAmountProps}
-            />
-
-            <ErrorMessage>{errors.amount?.message}</ErrorMessage>
-          </TransactionInputCard.Root>
+          <DelegationInput
+            amountError={errors.amount?.message}
+            delegatorInfo={delegatorInfo}
+            register={register}
+            setValue={setValue}
+            watch={watch}
+          />
         </div>
 
         <div className="flex flex-col justify-between gap-4 grow">
